controller/requirements.txt
flask==1.1.1
paho-mqtt==1.5.0
transitions==0.7.1
